requirements.txt
matplotlib==3.6.2
numpy==1.23.4
opencv_python==4.6.0.66
pandas==1.5.1
Pillow==9.2.0
scikit_learn==1.1.3
shap==0.41.0
timm==0.6.11
torch==1.12.1
torchmetrics==0.10.3
torchvision==0.13.1